A controlled vocabulary document, usually pointing to an .obo file

A controlled vocabulary document, usually pointing to an .obo file

Fields

name

Full name of the controlled vocabulary.

uri

Publicly accessible URI of the controlled vocabulary.

version

(optional) Version of the controlled vocabulary.

Examples

  MzQCcontrolledVocabulary$new(
    "Proteomics Standards Initiative Quality Control Ontology",
    "https://github.com/HUPO-PSI/psi-ms-CV/releases/download/v4.1.129/psi-ms.obo",
    "4.1.129")
#> Reference class object of class "MzQCcontrolledVocabulary"
#> Field "name":
#> [1] "Proteomics Standards Initiative Quality Control Ontology"
#> Field "uri":
#> [1] "https://github.com/HUPO-PSI/psi-ms-CV/releases/download/v4.1.129/psi-ms.obo"
#> Field "version":
#> [1] "4.1.129"