A controlled vocabulary parameter, as detailed in the OBO file

A controlled vocabulary parameter, as detailed in the OBO file

A controlled vocabulary parameter, as detailed in the OBO file

Public fields

accession

Accession number identifying the term within its controlled vocabulary.

name

Name of the controlled vocabulary term describing the parameter.

value

(optional) Value of the parameter.

description

(optional) Definition of the controlled vocabulary term.

Methods

Public methods

MzQCcvParameter$new()

MzQCcvParameter$isValid()

MzQCcvParameter$toJSON()

MzQCcvParameter$fromData()

MzQCcvParameter$clone()

Method new()

Constructor

Usage

MzQCcvParameter$new(
  accession = NA_character_,
  name = NA_character_,
  value = NA,
  description = NA_character_
)

Arguments

accession

String value for initialization of field accession

name

String value for initialization of field name

value

Optional value for initialization of field value

description

Optional string value for initialization of field description

Method isValid()

Verifies validity of the object

Usage

MzQCcvParameter$isValid(context = "MzQCcvParameter")

Arguments

context

Optional string describing location in mzQC structure that is used for more informative warning texts.

Method toJSON()

Creates JSON file from this object.

Usage

MzQCcvParameter$toJSON(...)

Arguments

...

Optional parameters for jsonlite:::asJSON()

Method fromData()

Sets data for this object from plain named lists of R objects

Usage

MzQCcvParameter$fromData(data, context = "MzQCcvParameter")

Arguments

data

A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'

context

Optional string describing location in mzQC structure that is used for more informative warning texts.

Method clone()

The objects of this class are cloneable with this method.

Usage

MzQCcvParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  MzQCcvParameter$new("MS:4000070",
                      "retention time acquisition range",
                      c(0.2959, 5969.8172))
#> <MzQCcvParameter>
#>   Public:
#>     accession: MS:4000070
#>     clone: function (deep = FALSE) 
#>     description: NA
#>     fromData: function (data, context = "MzQCcvParameter") 
#>     initialize: function (accession = NA_character_, name = NA_character_, value = NA, 
#>     isValid: function (context = "MzQCcvParameter") 
#>     name: retention time acquisition range
#>     self: MzQCcvParameter, R6
#>     toJSON: function (...) 
#>     value: 0.2959 5969.8172
  isValidMzQC(MzQCcvParameter$new("MS:0000000"))
#> Warning: MzQCcvParameter$self$name is NA/NULL
#> [1] FALSE