Root element of an mzQC document

Root element of an mzQC document

Root element of an mzQC document

Details

At least one of runQualities or setQualities MUST be present.

Public fields

version

Version of the mzQC format.

creationDate

Creation date of the mzQC file.

contactName

Name of the operator/creator of this mzQC file.

contactAddress

Contact address (mail/e-mail or phone)

description

Description and comments about the mzQC file contents.

runQualities

Array of MzQCrunQuality;

setQualities

Array of MzQCsetQuality

controlledVocabularies

Array of CV domains used (obo files)

Methods

Public methods

MzQCmzQC$new()

MzQCmzQC$isValid()

MzQCmzQC$toJSON()

MzQCmzQC$fromData()

Method new()

Constructor

Usage

MzQCmzQC$new(
  version = NA_character_,
  creationDate = MzQCDateTime$new(),
  contactName = NA_character_,
  contactAddress = NA_character_,
  description = NA_character_,
  runQualities = list(),
  setQualities = list(),
  controlledVocabularies = list()
)

Arguments

version

String value for initialization of field version

creationDate

MzQCDateTime value for initialization of field creationDate

contactName

Optional string value for initialization of field contactName

contactAddress

Optional string value for initialization of field contactAddress

description

Optional string value for initialization of field description

runQualities

Optional list value for initialization of field runQualities

setQualities

Optional list value for initialization of field setQualities

controlledVocabularies

Optional list value for initialization of field controlledVocabularies

Method isValid()

Verifies validity of the object

Usage

MzQCmzQC$isValid(context = "MzQCmzQC")

Arguments

context

Optional string describing location in mzQC structure that is used for more informative warning texts.

Method toJSON()

Creates JSON file from this object.

Usage

MzQCmzQC$toJSON(...)

Arguments

...

Optional parameters for jsonlite:::asJSON()

Method fromData()

Sets data for this object from plain named lists of R objects

Usage

MzQCmzQC$fromData(data, context = "MzQCmzQC")

Arguments

data

A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'

context

Optional string describing location in mzQC structure that is used for more informative warning texts.