The central class to store QC information

The central class to store QC information

The central class to store QC information

Public fields

accession

Accession number identifying the term within its controlled vocabulary.

name

Name of the controlled vocabulary element describing the metric.

description

(optional) Definition of the controlled vocabulary term.

value

(optional) Value of the metric (single value, n-tuple, table, matrix). The structure is not checked by our mzQC implementation and must be handled by the caller, see toQCMetric.

unit

(optional) Array of unit(s), stored as MzQcvParameter

Methods

Public methods

MzQCqualityMetric$new()

MzQCqualityMetric$isValid()

MzQCqualityMetric$toJSON()

MzQCqualityMetric$fromData()

Method new()

Constructor

Usage

MzQCqualityMetric$new(
  accession = NA_character_,
  name = NA_character_,
  description = NA_character_,
  value = NA,
  unit = list()
)

Arguments

accession

String value for initialization of field accession

name

String value for initialization of field name

description

Optional string value for initialization of field description

value

Optional value for initialization of field value

unit

Optional unit value for initialization of field unit

Method isValid()

Verifies validity of the object

Usage

MzQCqualityMetric$isValid(context = "MzQCqualityMetric")

Arguments

context

Optional string describing location in mzQC structure that is used for more informative warning texts.

Method toJSON()

Creates JSON file from this object.

Usage

MzQCqualityMetric$toJSON(...)

Arguments

...

Optional parameters for jsonlite:::asJSON()

Method fromData()

Sets data for this object from plain named lists of R objects

Usage

MzQCqualityMetric$fromData(data, context = "MzQCqualityMetric")

Arguments

data

A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'

context

Optional string describing location in mzQC structure that is used for more informative warning texts.