Allow conversion of plain named lists of R objects (from jSON) to mzQC objects

Allow conversion of plain named lists of R objects (from jSON) to mzQC objects

Usage

fromDatatoMzQC(mzqc_class, data, context = NULL)

Arguments

mzqc_class

Prototype of the class to convert 'data' into

data

A list of: A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'

context

A trace through the mzQC object tree to aid users in case of errors

Examples

    data = rmzqc::MzQCcvParameter$new("acc", "myName", "value")
    data_recovered = rmzqc::fromDatatoMzQC(rmzqc::MzQCcvParameter,
                         list(jsonlite::fromJSON(jsonlite::toJSON(data))))